Setup & Config
Configured vectorizer, docker envir, template for RSD data mode in Weviate
