+++ docker-compose.yml
@@ -1,0 +1,1 @@
+{"error":"missing required parameter \"weaviate_volume\""}
